+++ src/main/java/gui/controllers/AnalysisResult.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.controllers;
 
 
 import detail.compute.DescriptiveStats;
