+++ src/test/java/gui/controllers/JStatGuiIndexTest.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.controllers;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
